Use OA hash as document id
The OA hash is calculated from the data in each row, so rows with
identical data have identical hashes. This is nice for two reasons:

1.) It means bit-for-bit duplicate records will no longer be inserted
into Elasticsearch multiple times. Instead one document will be updated
several times. This is not quite as efficient as, and is no substitute
for, actually removing duplicate lines from OA files before importing as
a preprocessing step, but it's still nice.

2.) Our current ID scheme simply uses a counter starting at zero, so you
can only find the record if you're sure you have the same file, and want
to count line numbers. The hash will be the same, assuming the data is
the same, even if the file being looked at is newer or older than the
one that was used to import.