Remove section on deduplication
As part of https://github.com/pelias/address-deduplicator/pull/22 it
makes sense to officially deprecate the deduplicator. Removing this
entire section of the readme should help avoid confusion.

The deduplication option in the configuration section is now at the
bottom (to be less prominent), and a link to the readme which will soon
have deprecation information is now provided.