feat(import): add missingFilesAreFatal option
This adds a config option similar to the one in the [Who's on First importer](https://github.com/pelias/whosonfirst#configuration),
which will allow users of this importer to decide if a missing file
should count as a fatal error or be skipped gracefully.

Previously, this importer would import files until it got to one that
doesn't exist, and then exit with an unclear error.

Now, the importer either quits right away (if `missingFilesAreFatal` is
true), or emits a warning message and continues with the other files.
Gracefully skipping missing files is now the default.

Fixes https://github.com/pelias/openaddresses/issues/311