feat(logger): Use dbclient name functionality
This allows distinguishing which logger output lines correspond to which
importer during an import.

Unlike other importers, this PR is slightly more complicated as we have
to support unique names for each of what might be several parallel
sub-importers.

Dependency injection for `dbclient` was removed in favor of using
`proxyquire` for testing to facilitate this.

Connects https://github.com/pelias/dbclient/issues/101