feat(logs): Remove bad record count verbose logger
This log has always been a pain. It's on a simple 10 second timer so it
will print forever, even if nothing has changed.

It's measuring the number of records that were invalid, which is not a
particularly notable thing. There are invalid records in OA data all the
time.

Overall, we don't really need this and it's just cluttering up other,
more important, log output.